+++ ChangeLog.md
@@ -3,6 +3,13 @@ XP Framework Core ChangeLog
 
 ## ?.?.? / ????-??-??
 
+### Features
+
+* Added new method `lang.Type::resolve()` to correctly resolve types in
+  a given context. Adds support for `array<self>` and others for return
+  and parameter types.
+  (@thekid)
+
 ## 10.3.3 / 2020-11-22
 
 ### Bugfixes